requirements.txt
agno==1.3.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
auth0-python==4.8.1
backoff==2.2.1
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.105.0
cryptography==44.0.2
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
durationpy==0.9
ecdsa==0.19.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.104.1
filelock==3.17.0
flatbuffers==25.2.10
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
gotrue==2.12.0
grpcio==1.71.0
gspread==6.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
iniconfig==2.0.0
instructor==1.7.3
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
json5==0.10.0
json_repair==0.39.1
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.0
kubernetes==32.0.1
langchain==0.3.20
langchain-anthropic==0.3.10
langchain-community==0.3.19
langchain-core==0.3.51
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langsmith==0.3.13
litellm==1.60.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.30.0
networkx==3.4.2
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.65.5
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==2.1.4
parso==0.8.4
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==10.4.0
plotly==5.18.0
pluggy==1.5.0
postgrest==1.0.1
posthog==3.19.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.25.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.6
pytz==2025.1
pyvis==0.3.2
PyYAML==6.0.2
realtime==2.4.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.27.0
storage3==0.11.3
streamlit==1.29.0
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uv==0.6.5
uvicorn==0.24.0
uvloop==0.21.0
validators==0.34.0
vine==5.1.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
web3storage==0.1.1
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
